Line up x belt
Shift the position of the motor based on the pulley and the idler based on the bearing so that the belt lines up with the x carriage. Supports 623 and 624 bearings with printed bearing guides. Does not support 608 bearings for idler. 635 bearings must be doubled up and used without a bearing guide.